refactor data transformer file name to class name
