+++ deepparse/converter/__init__.py
@@ -1,4 +1,4 @@
 # pylint: disable=wildcard-import
 from .data_padding import *
 from .target_converter import *
-from .data_transformer import *
+from .data_transform import *
